C++ time & timeless support (#3838)
### What

* Fixes #3394

Introduces time methods to C/C++, timeless logging and ports the clock
example.

Originally, I planned to expose a Timeline struct type to C but this
actually created more surface for errors to occur, so I instead expose
the three time setter for seconds/nanos/sequences separately, making the
code much easier to reason with.

<img width="1421" alt="image"
src="https://github.com/rerun-io/rerun/assets/1220815/a21ab48e-7ced-4dc2-85bc-77e8d99eb452">


### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] I have tested [demo.rerun.io](https://demo.rerun.io/pr/3838) (if
applicable)
* [x] The PR title and labels are set such as to maximize their
usefulness for the next release's CHANGELOG

- [PR Build Summary](https://build.rerun.io/pr/3838)
- [Docs
preview](https://rerun.io/preview/81bd132335ca725ae0c2a1fb2c42156f5b1986d3/docs)
<!--DOCS-PREVIEW-->
- [Examples
preview](https://rerun.io/preview/81bd132335ca725ae0c2a1fb2c42156f5b1986d3/examples)
<!--EXAMPLES-PREVIEW-->
- [Recent benchmark results](https://ref.rerun.io/dev/bench/)
- [Wasm size tracking](https://ref.rerun.io/dev/sizes/)

+++ docs/code-examples/clear_recursive.cpp
@@ -29,7 +29,7 @@ int main() {
     };
 
     // Log a handful of arrows.
-    for (int i = 0; i < vectors.size(); ++i) {
+    for (size_t i = 0; i < vectors.size(); ++i) {
         auto entity_path = "arrows/" + std::to_string(i);
         rec.log(
             entity_path.c_str(),
